requirements.txt
flask
python-dotenv
google-generativeai
flask-sqlalchemy
flask-mail
